Adds AuthnStatement claims for non-password authentication methods
The claims forming the AuthnStatement were only added when the authentication method is "password". An external authentication in IdentityServer3 sets the authentication method to "external", thus no AuthnStatement was generated and the service provider receiving the response could not create a session. This commit will cause the claims for AuthnStatement to be added always, if the authentication method differs from "password", the value for the AuthenticationMethod claim is AuthenticationMethods.Unspecified.